requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
google-cloud-texttospeech==2.14.1
python-dotenv==1.0.0
openai==1.3.0
google-generativeai==0.3.0
SpeechRecognition==3.10.0
Pillow==10.1.0
pydantic==2.5.2
